requirements.txt
certifi==2020.12.5
MouseInfo==0.1.3
Pillow==8.1.2
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.7
python3-xlib==0.15
PyTweening==1.0.3
